refactor: remove disused code and move componentsX
