--- src/vtk/CustomDistanceWidget/state.js
@@ -1,26 +1,0 @@
-import vtkStateBuilder from '@kitware/vtk.js/Widgets/Core/StateBuilder';
-
-export default function generateState() {
-  return vtkStateBuilder
-    .createBuilder()
-    .addStateFromMixin({
-      labels: ['moveHandle'],
-      mixins: ['origin', 'color', 'scale1', 'visible'],
-      name: 'moveHandle',
-      initialValues: {
-        scale1: 30,
-        origin: [0, 0, 0],
-        visible: false,
-      },
-    })
-    .addDynamicMixinState({
-      labels: ['handles'],
-      mixins: ['origin', 'color', 'scale1'],
-      name: 'handle',
-      initialValues: {
-        scale1: 30,
-        origin: [0, 0, 0],
-      },
-    })
-    .build();
-}
